requirements.txt
ipykernel
notebook

mlflow>=2.8.0
transformers>=4.36.0
torch>=2.1.0
datasets>=2.14.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pytest>=7.4.0
jaxtyping
transformer-lens
transformers_stream_generator
einops

deepeval
tqdm
nanogcg

colorama
wandb
matplotlib
seaborn

google-genai
